Fixes for CGE and Miscellaneous Fixes
- Fixed up importing of ComboGestureExpressions and the reimporting method for Hai's assets.
- Added a Social button for my website: https://kebin.dev.
- Renamed Donate MenuItem Folder to Support.

+++ Editor/Scripts/Items/100-330_Common/330_CGE.cs
@@ -17,8 +17,40 @@ public partial class kebinImports : MonoBehaviour
         bool refreshRequired = false;
         if (Directory.Exists(Application.dataPath + @"/Hai"))
         {
-            Directory.Delete(Application.dataPath + @"/Hai", true);
-            refreshRequired = true;
+            if (Directory.Exists(Application.dataPath + @"/Hai/AnimationViewer"))
+            {
+                Directory.Delete(Application.dataPath + @"/Hai/AnimationViewer", true);
+                refreshRequired = true;
+            }
+            if (File.Exists(Application.dataPath + @"/Hai/AnimationViewer.meta"))
+            {
+                File.Delete(Application.dataPath + @"/Hai/AnimationViewer.meta");
+                refreshRequired = true;
+            }
+            if (Directory.Exists(Application.dataPath + @"/Hai/ComboGesture"))
+            {
+                Directory.Delete(Application.dataPath + @"/Hai/ComboGesture", true);
+                refreshRequired = true;
+            }
+            if (File.Exists(Application.dataPath + @"/Hai/ComboGesture.meta"))
+            {
+                File.Delete(Application.dataPath + @"/Hai/ComboGesture.meta");
+                refreshRequired = true;
+            }
+            if (Directory.Exists(Application.dataPath + @"/Hai/VisualExpressionsEditor"))
+            {
+                Directory.Delete(Application.dataPath + @"/Hai/VisualExpressionsEditor", true);
+                refreshRequired = true;
+            }
+            if (File.Exists(Application.dataPath + @"/Hai/VisualExpressionsEditor.meta"))
+            {
+                File.Delete(Application.dataPath + @"/Hai/VisualExpressionsEditor.meta");
+                refreshRequired = true;
+            }
+            if (Directory.GetFiles(Application.dataPath + @"/Hai").Length == 0 && Directory.GetDirectories(Application.dataPath + @"/Hai").Length == 0)
+            {
+                Directory.Delete(Application.dataPath + @"/Hai", true);
+            }
         }
         if (File.Exists(Application.dataPath + @"/Hai.meta"))
         {
@@ -30,6 +62,6 @@ public partial class kebinImports : MonoBehaviour
 
         // Import Asset
         string repoLink = "https://api.github.com/repos/hai-vr/combo-gesture-expressions-av3/releases/latest";
-        ImportAsset(JSON.Parse(Task.Run(() => ModHttpClient.DownloadString(client,repoLink)).Result)["assets"][0]["browser_download_url"]);
+        ImportAsset(JSON.Parse(Task.Run(() => ModHttpClient.DownloadString(client, repoLink)).Result)["assets"][0]["browser_download_url"], false, true);
     }
 }
